Correcciones
git-svn-id: https://draco.esm.cpqcorp.net/svn/DEG_ORANGE/branches/delivery@155 1b308f40-575c-0510-bb77-a5e8b53bb2f9

+++ templates_tobuild/var/opt/SIU/scripts/jobs/kpis/job-service-kpis/sh/job-service-kpis.sh
@@ -13,6 +13,18 @@ function process
 {
   logDebug "Executing function 'process'"
 
+  ENABLE_SIMULATION="$(getConfigParam GENERAL ENABLE_SIMULATION)"
+  if [ $? -lt 0 ] || [ -z "${ENABLE_SIMULATION}" ]
+  then
+    logWarning "Unable to get parameter 'ENABLE_SIMULATION' in section 'GENERAL'"
+  else
+    if [ "${ENABLE_SIMULATION}" != "TRUE"]
+    then
+      ENABLE_SIMULATION=FALSE
+    fi
+  fi
+  logDebug "ENABLE_SIMULATION = ${ENABLE_SIMULATION}"
+
   /var/opt/<%SIU_INSTANCE%>/KPI/scripts/KPIReport.sh
   if [ $? -ne 0 ]
   then
@@ -40,8 +52,38 @@ function process
     NEW_FILENAME="hpedeg-service_kpis_request_types-${HOSTNAME_OSS}-"$(echo ${FILENAME} | cut -d "_" -f 2 | cut -c 1-12)
     logDebug "NEW_FILENAME = ${NEW_FILENAME}"
 
-    cp ${FILEPATH} /var/opt/<%SIU_INSTANCE%>/KPI/OSS/${NEW_FILENAME}
-    if [ $? -ne 0 ]
+    if [ "${ENABLE_SIMULATION}" = "FALSE" ]
+    then
+      cp ${FILEPATH} /var/opt/<%SIU_INSTANCE%>/KPI/OSS/${NEW_FILENAME}
+      RESULT=$?
+    else
+      cat ${FILEPATH} | awk -F \| -v seed=${RANDOM} '
+      function randint(n)
+      {
+        return int(n * rand());
+      }
+      BEGIN
+      {
+        srand(seed);
+      }
+      {
+        prefix = $1"|"$2"|"$3"|"$4"|"$5"|"$6"|"$7;
+        action = $5;
+
+        if ((action == "DER") || (action == "getAuthentication") || (action == "postChallenge") || (action == "getPhoneNumber") || (action == "getEntitlement") || (action == "enablePushNotificationEntitlement") || (action == "getPhoneServicesAccountStatus") || (action == "updatePushToken")) {
+          success = randint(1000);
+          unsuccess = int(success * rand());
+          average_time = sprintf("%.3f", rand());
+          print prefix"|"success"|"unsuccess"|"average_time;
+        }
+        else {
+          print $0;
+        }
+      }' > /var/opt/<%SIU_INSTANCE%>/KPI/OSS/${NEW_FILENAME}
+      RESULT=$?
+    fi
+
+    if [ ${RESULT} -ne 0 ]
     then
       logError "Command 'cp ${FILEPATH} /var/opt/<%SIU_INSTANCE%>/KPI/OSS/${NEW_FILENAME}' failed"
       RETURN_CODE=1
@@ -76,6 +118,7 @@ function process
 
     NEW_FILENAME="hpedeg-service_kpis_return_codes-${HOSTNAME_OSS}-"$(echo ${FILENAME} | cut -d "_" -f 2 | cut -c 1-12)
     logDebug "NEW_FILENAME = ${NEW_FILENAME}"
+
 
     cp ${FILEPATH} /var/opt/<%SIU_INSTANCE%>/KPI/OSS/${NEW_FILENAME}
     if [ $? -ne 0 ]
